pyproject.toml
[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[project]
name = "teamvault"
description = "Keep your passwords behind the firewall"
readme = "README.md"
license = "GPL-3.0-only"
license-files = ["LICENSE"]
authors = [{ name = "Seibert Group GmbH" }]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Office/Business",
    "Topic :: Security",
]
keywords = ["password", "safe", "manager", "sharing"]
dependencies = [
    "cryptography~=47.0.0",
    "django-auth-ldap~=5.3.0",
    "django-bootstrap5==26.2",
    "django-filter==25.2",
    "django-htmx~=1.27.0",
    "django-test-migrations>=1.5.0",
    "django-webpack-loader~=3.2.3",
    "django~=6.0.2",
    "djangorestframework~=3.17.1",
    "gunicorn~=25.3.0",
    "hashids~=1.3.1",
    "pyotp~=2.9",
    "huey~=3.0.0",
    "psycopg~=3.3.2",
    "requests~=2.32",
    "social-auth-app-django~=5.8.0",
    "whitenoise[brotli]~=6.12.0",
]

version = "0.13.0"  # automatically changed via "uv version"

[dependency-groups]
dev = [
    "django-stubs~=6.0.3",
    "djangorestframework-stubs~=3.16.0",
    "faker",
    "playwright>=1.58.0",
    "ruff>=0.14.6",
]

[project.scripts]
teamvault = "teamvault.cli:main"

[project.urls]
Source = "https://github.com/seibert-media/teamvault"

[tool.uv]
package = true
exclude-newer = "7 days"

[tool.uv.build-backend]
module-root = ""
source-include = [
    "CHANGELOG.md",
    "MANIFEST.in",
]

[tool.uv.sources]
teamvault = { workspace = true }

[tool.ruff]
src = ["teamvault"]
target-version = "py312"
line-length = 120
indent-width = 4
preview = true
extend-exclude = [
    "teamvault/apps/*/migrations"
]

[tool.ruff.format]
# Like Black, use single quotes for strings.
quote-style = "single"

[tool.ruff.lint]
select = [
    # Note: Preferably, we'd use the deactivated rules below, too. Once there's time, let's try to fix them one by one.
    #
    "ARG",   # flake8-unused-arguments - when enabling, make sure you keep the same signature of parent methods
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "DJ",      # flake-django
    "E",       # pycodestyle errors
    "F",       # pyflakes
    "G",     # flake8-logging-format
    "FURB",  # refurb  - depends on enabled preview option
    "I",       # isort
    "INT",     # flake8-gettext
    "ISC",     # flake8-implicit-str-concat
    "PERF",    # perflint
    "PLE",     # pylint errors
    "PLR",  # pylint refactor errors
    "PTH",   # flake8-use-pathlib
    "RUF100",  # ruff / unused-noqa
    "SIM",   # flake8-simplify
    "TCH",     # flake8-type-checking
    "UP",      # pyupgrade
    "W",       # pycodestyle warnings
]
ignore = [
    "DJ008",   # django-model-without-dunder-str - Often, we don't care about these
    "PERF401", # manual-list-comprehension - Forcing us to use list comprehensions with dicts inside them is weird.
    "UP015",   # redundant-open-modes - We like to have the mode explicitly set in open() calls
]

[tool.ruff.lint.isort]
combine-as-imports = true
no-lines-before = ["local-folder"]
relative-imports-order = "closest-to-furthest"

[tool.ruff.lint.mccabe]
max-complexity = 13

[tool.ruff.lint.pylint]
max-args = 15  # default is 5
max-branches = 28  # default is 12
max-returns = 13  # default is 6
max-statements = 134  # default is 50
